backend/requirements.txt
django
mysqlclient
django-mysql
djangorestframework
pytest
py3-validate-email
gym
numpy
pandas
networkx
requests
tqdm
urllib3
drf-yasg
sendgrid
django-storages
boto3
bcrypt
faker